Return *ConnectionError when connection has faulted
Errors reading from/writing to the underlying net.Conn were being
propagated to the sender/receiver APIs, making them hard to identify.
Now, when the connection is no longer functional, a ConnectionError
instance is returned.  If the connection died due to a read/write error,
the ConnectionError will contain this information.
Removed ErrConnClosed as it's too simplistic and can't capture the
reason why the connection was closed.