Fix tracking of link credit during receives
Links can receive frames in two places (ie call muxHandleFrame).  Either
directly from link.mux or from link.muxFlow.  In both cases, once a
frame has been received, we need to reevaluate the link credit to
determine if we need to send a flow frame or pause the link.
Previously, this was not happening for frames received in muxFlow() via
doFlow() which was the reason for the sporadic test failures.