Add retries when writing to net.Conn
Retry with exponential back-off when writing to the network fails.  The
initial delay time is 100ms.
Capture write errors when sending the close performative.
Fixed up some misbehaving tests that were uncovered due to the above
changes.
Updated Go version to 1.13 as there are some features used that first
appeared in that version.