Remove link.Paused
Asking the link mux to evaluate its state is fine in any case.
There was a corner case where this would incorrectly be set.  If a
transfer frame is receiving from link.muxFlow() the link wouldn't be
marked as paused even though the credit was consumed.  Fixing this is
complex and doesn't make sense for an unnecessary feature.