Fix potential hang when closing a receiver
Select on r.l.close in muxReceive instead of r.l.done.  The latter is
only closed when the mux exits, and since muxReceive is called from the
mux, this case will never trigger.
Add a debug build for selecting on r.messages that will panic if it's
full.  This indicates a flow control bug.