Replace messages channel with a queue (#244)
* Replace messages channel with a queue

The Receiver.messages channel placed an arbitrary restriction on the
amount of link credit that could be issued to a Receiver.  Once the
Receiver has been created, it can never be issued credit that exceeds
the size of the channel.
In addition, if the Receiver were to receive messages that exceeds the
size of the channel, due to flow control bugs or other, the writes to
the channel could block leading to hangs.
The channel has been replaced with a segmented FIFO queue.  While this
in theory could allow for unbounded growth, the reality is that the
total size can never be greater than that of the Session's incoming
window.

* use Ring from standard library

* switch to require.Same for pointer comparisons

fixed issue in last seg check and added another test

* refine and add another test

* size message queue to the incoming window

* add Holder[T] to abstract syncronized access