Improved logging diagnostics
Include address of Conn/Session/Sender/Receiver when writing log
entries.  This helps when a process contains multiple instances.
Include the link name in the error string when a link is forcibly
closed.
Include the channel number in the error string when a session is
forcibly closed.
Add log entries when session/link is forcibly closed.
Add log entry for muxing from Session to Conn.