Make writes deterministic
Added Conn.WriteDeadline to control the write deadline when writing to
net.Conn.  The default value is 30 seconds.  This can be overridden by
APIs that take a context and the context has a deadline/timeout.
For APIs that directly send frames, wait for the frame to be written to
the underlying net.Conn before considering the write a success.